refactor runtime bridge helper modules
