+++ .sampo/changesets/issue-361-runtime-bridge-helpers.md
@@ -1,0 +1,5 @@
+---
+npm/wp-typia: patch
+---
+
+Split runtime bridge output and sync helpers into focused modules without changing the public bridge surface.
